make sure user is a bootcamp owner or an admin to update or delete a bootcamp

+++ controllers/bootcamps.js
@@ -15,7 +15,9 @@ exports.getBootcamps = asyncHandler(async (req, res, next) => {
 // @route    GET /api/v1/bootcamps
 // @access   Public
 exports.getBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findById(req.params.id).populate("courses");
+  const bootcamp = await Bootcamp.findById(req.params.id)
+    .populate("courses")
+    .populate("user");
 
   if (!bootcamp) {
     return next(
@@ -58,16 +60,28 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route    PUT /api/v1/bootcamps/:id
 // @access   Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
+
+  // make sure that the user is a bootcamp owner or an admin
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.params.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
+
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
 
   res.status(200).json({ success: true, data: bootcamp });
 });
@@ -79,6 +93,16 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
   const bootcamp = await Bootcamp.findById(req.params.id);
 
   if (bootcamp) {
+    // make sure that the user is a bootcamp owner or an admin
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.params.id} is not authorized to update this bootcamp`,
+          401
+        )
+      );
+    }
+
     await Bootcamp.deleteOne({ _id: bootcamp._id });
 
     // For cascading delete
